refactor: Move fetchCollection and UseCollectionSchema to collectionOptions module
Relocate collection-related schema and fetch function from useCollection to collectionOptions for better code organization and modularity
